project_task_employee: Refactoring code
