+++ project_task_employee/models/__init__.py
@@ -1,2 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 from . import project
+from . import project_task
